Revision 4: Address FE review comments
- Rename showHome → showOlliePage (only controls Ollie visibility now)
- Replace Button+onClick with Link for nav buttons and activity items
- Move formatRelativeDateTime to @/lib/date.ts alongside other date utils
