+++ apps/opik-frontend/src/v2/layout/SideBar/SideBarMenuItems.tsx
@@ -33,7 +33,7 @@ const SideBarMenuItems: React.FC<SideBarMenuItemsProps> = ({ expanded }) => {
     canViewDashboards,
     canUsePlayground,
     canViewOptimizationRuns,
-    showHome: !!AssistantSidebar,
+    showOlliePage: !!AssistantSidebar,
   });
 
   const renderItems = (items: MenuItem[]) => {
